blender/requirements.txt
# Core WebSocket and networking
websocket-client>=1.6.0,<2.0.0
requests>=2.25.0,<3.0.0

# HTTP and SSL support
certifi>=2021.0.0
urllib3>=1.26.0,<3.0.0
charset-normalizer>=2.0.0,<4.0.0
idna>=2.10,<4.0.0

# Development and packaging
setuptools>=50.0.0
